Finished viewOrder + fixed method name in deleteOrder + moved apis to appropriate directory
